% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/notifyFunctions.R
\name{jobNotify}
\alias{jobNotify}
\title{notifies job status}
\usage{
jobNotify(jobRun = function() { }, jobID, userID = "", addNow = TRUE,
  testTweet = FALSE)
}
\arguments{
\item{jobRun}{the function call to evaluate (default is empty function)}

\item{jobID}{text string to associate with job call}

\item{userID}{the users to notify (default is "")}

\item{addNow}{add current date-time to tweet text}

\item{testTweet}{if TRUE, print the tweets instead of actually tweeting them}
}
\description{
sends tweets to users about the job status
}

